launch emitters/atdome-sim together and reorganize folders
