+++ simulator/atdome/simulator.py
@@ -21,5 +21,5 @@ async def main():
         await asyncio.sleep(10)
 
 if __name__ == '__main__':
-    print('starting loop')
+    print('starting atdome loop')
     asyncio.get_event_loop().run_until_complete(main())
