+++ simulator/emitters/emitter_example_dict.py
