+++ simulator/emitters/simulator.py
@@ -2,14 +2,14 @@ import asyncio
 import importlib
 import threading
 from lsst.ts import salobj
-from emitter import emit_forever
-from event_emitter import emit_forever as emit_event_forever
+from .emitter import emit_forever
+from .event_emitter import emit_forever as emit_event_forever
 
 
 def add_controller_in_thread(sal_lib, loop, index):
     asyncio.set_event_loop(loop)
     controller = create_controller(sal_lib, index)
-    launch_emitters_forever(controller)
+    launch_emitters_forever(loop, controller)
 
 
 def create_controller(sallib, index):
@@ -19,7 +19,7 @@ def create_controller(sallib, index):
     return controller
 
 
-def launch_emitters_forever(controller):
+def launch_emitters_forever(loop, controller):
     """
         Launches an emitter that fills the data to be read
         later in the salobj remote
@@ -31,15 +31,14 @@ def launch_emitters_forever(controller):
     t2.start()
 
 
-def run_evt_loop(loop):
-    loop.run_forever()
+# def run_evt_loop(loop):
+#     loop.run_forever()
 
 
-if __name__ == '__main__':
+async def main(loop):
     print('--main--')
-    loop = asyncio.get_event_loop()
-    t = threading.Thread(target=run_evt_loop, args=(loop,))
-    t.start()
+    # t = threading.Thread(target=run_evt_loop, args=(loop,))
+    # t.start()
 
     sal_lib_param_list = [line.rstrip('\n') for line in open('/usr/src/love/sallibs.config')]
     for i in range(len(sal_lib_param_list)):
